# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fluxa API
  version: 1.0.0
  description: 'Fluxa is a multi-currency digital asset platform built on Stellar. It provides

    wallets, instant transfers, FX conversions, scheduled recurring payments,

    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.


    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.

    '
  contact:
    name: Fluxa Engineering
    email: engineering@fluxa.io
    url: https://github.com/fluxa/fluxa
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:3000
  description: Local development
- url: https://api.fluxa.io
  description: Production
...
security:
- BearerAuth: []
paths:
...
  /v1/webhooks/secret:
    get:
      operationId: listWebhookSigningSecrets
      summary: Inspect webhook signing secret versions
      description: Returns version metadata only. Secret material is never returned by this endpoint.
      tags:
      - Webhooks
      responses:
        '200':
          description: Signing-secret metadata for the authenticated tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSigningSecretsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/webhooks/secret/rotate:
    post:
      operationId: rotateWebhookSigningSecret
      summary: Rotate the tenant webhook signing secret
      description: Generates a new secret and returns its plaintext once. A bounded overlap lets in-flight deliveries signed with the prior version remain verifiable.
      tags:
      - Webhooks
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RotateWebhookSigningSecretRequest'
      responses:
        '200':
          description: New signing secret, disclosed once for this idempotent operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotateWebhookSigningSecretResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Another request with the same idempotency key is still processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/webhooks:
    post:
      operationId: registerWebhook
      summary: Register a webhook endpoint
      description: 'Registers a new webhook endpoint. A signing secret is generated and

        returned — it will be used to sign webhook payloads via HMAC-SHA256.

        '
      tags:
      - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterWebhookRequest'
            example:
              url: https://acme.com/hooks/fluxa
              events:
              - transfer.initiated
              - transfer.settled
      responses:
        '201':
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpointResponse'
              example:
                id: wh_abc123
                url: https://acme.com/hooks/fluxa
                secret: whsec_abc123def456
                events:
                - transfer.initiated
                - transfer.settled
                active: true
                created_at: '2026-01-15T10:00:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      operationId: listWebhooks
      summary: List webhook endpoints
      description: Returns all registered webhook endpoints for the tenant.
      tags:
      - Webhooks
      responses:
        '200':
          description: List of webhook endpoints
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoints:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookEndpointSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/webhooks/{id}:
    delete:
      operationId: deleteWebhook
      summary: Delete a webhook endpoint
      description: Permanently removes the webhook endpoint.
      tags:
      - Webhooks
      parameters:
      - $ref: '#/components/parameters/WebhookID'
      responses:
        '204':
          description: Webhook deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: 'Pass a tenant API key (`sk_live_...` or `sk_test_...`) or JWT bearer token.

        In Swagger UI, enter your token in the Authorize dialog (the ''Bearer '' prefix is added automatically).

        '
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Pass an API key as `Bearer <key>` in the `Authorization` header.

        '
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: admin token
      description: Administrative bearer token.
  parameters:
...
    WebhookID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Webhook endpoint ID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: invalid request body
    Unauthorized:
      description: Unauthorized — missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: missing or invalid authorization header
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: resource not found
    Forbidden:
      description: Forbidden — admin access required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: admin access required
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: an unexpected error occurred
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              enum:
              - BAD_REQUEST
              - NOT_FOUND
              - UNAUTHORIZED
              - FORBIDDEN
              - INTERNAL_ERROR
              - QUOTE_EXPIRED
              - QUOTE_ALREADY_USED
              - REVIEW_ALREADY_DECIDED
              - CLAIMABLE_BALANCE_NOT_PENDING
              - REQUEST_IN_PROGRESS
              - IDEMPOTENCY_KEY_REUSED_WITH_DIFFERENT_BODY
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
...
    WebhookSigningSecretMetadata:
      type: object
      required:
      - key_id
      - created_at
      - activated_at
      - status
      properties:
        key_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        activated_at:
          type: string
          format: date-time
        retired_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [active, overlapping, retired]
    WebhookSigningSecretsResponse:
      type: object
      required: [secrets]
      properties:
        secrets:
          type: array
          items:
            $ref: '#/components/schemas/WebhookSigningSecretMetadata'
    RotateWebhookSigningSecretRequest:
      type: object
      properties:
        overlap_window_seconds:
          type: integer
          minimum: 0
          maximum: 86400
          default: 300
    RotateWebhookSigningSecretResponse:
      allOf:
      - $ref: '#/components/schemas/WebhookSigningSecretMetadata'
      - type: object
        required: [secret]
        properties:
          secret:
            type: string
            description: Plaintext signing secret; disclosed only by this rotation response.
          overlap_window_seconds:
            type: integer
    RegisterWebhookRequest:
      type: object
      required:
      - url
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
            - transfer.initiated
            - transfer.settled
            - transfer.failed
            - wallet.funded
            - conversion.completed
          description: 'Event types to subscribe to. Omit or pass an empty array to

            receive all events.

            '
    WebhookEndpointResponse:
      type: object
      description: Returned on create — includes the signing secret
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        secret:
          type: string
          description: HMAC-SHA256 signing secret (shown once)
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
    WebhookEndpointSummary:
      type: object
      description: Returned on list — secret is never included
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
...